Restructure amenities to be part of existing CRUD operations
- Remove separate amenities commands/entities as requested
- Integrate deployment cell amenities functionality into existing deployment cell CRUD operations (check-drift, sync, apply)
- Move organization amenities to service provider organization CRUD operations (init-amenities, update-amenities)
- Move amenities models into existing deployment cell and service models
- Move data access functions into existing hostcluster and service data access files
- Remove organization-id flag requirement - comes from credentials automatically
- Create sample-amenities.yaml file showing configuration format
- Update command help and documentation to reflect new structure
