+++ cmd/service/service.go
@@ -17,6 +17,8 @@ func init() {
 	Cmd.AddCommand(describeCmd)
 	Cmd.AddCommand(deleteCmd)
 	Cmd.AddCommand(listCmd)
+	Cmd.AddCommand(amenitiesInitCmd)
+	Cmd.AddCommand(amenitiesUpdateCmd)
 }
 
 func run(cmd *cobra.Command, args []string) {
