--- internal/dataaccess/amenities.go
@@ -1,221 +1,0 @@
-package dataaccess
-
-import (
-	"context"
-	"fmt"
-	"net/http"
-	"time"
-
-	"github.com/omnistrate-oss/omnistrate-ctl/internal/model"
-	openapiclientfleet "github.com/omnistrate-oss/omnistrate-sdk-go/fleet"
-)
-
-// InitializeOrganizationAmenitiesConfiguration initializes the organization-level amenities configuration
-// This is a placeholder implementation - the actual API endpoint may not exist yet
-func InitializeOrganizationAmenitiesConfiguration(ctx context.Context, token string, organizationID string, environment string, configTemplate map[string]interface{}) (*model.AmenitiesConfiguration, error) {
-	// TODO: Replace with actual API call once backend is available
-	// For now, return a mock response to enable testing of the CLI interface
-	
-	config := &model.AmenitiesConfiguration{
-		OrganizationID:        organizationID,
-		Environment:          environment,
-		ConfigurationTemplate: configTemplate,
-		Version:              "1.0.0",
-	}
-	
-	return config, nil
-}
-
-// UpdateOrganizationAmenitiesConfiguration updates the amenities configuration for a target environment
-// This is a placeholder implementation - the actual API endpoint may not exist yet
-func UpdateOrganizationAmenitiesConfiguration(ctx context.Context, token string, organizationID string, environment string, configTemplate map[string]interface{}) (*model.AmenitiesConfiguration, error) {
-	// TODO: Replace with actual API call once backend is available
-	
-	config := &model.AmenitiesConfiguration{
-		OrganizationID:        organizationID,
-		Environment:          environment,
-		ConfigurationTemplate: configTemplate,
-		Version:              "1.1.0",
-	}
-	
-	return config, nil
-}
-
-// GetOrganizationAmenitiesConfiguration retrieves the amenities configuration for an organization and environment
-// This is a placeholder implementation - the actual API endpoint may not exist yet
-func GetOrganizationAmenitiesConfiguration(ctx context.Context, token string, organizationID string, environment string) (*model.AmenitiesConfiguration, error) {
-	// TODO: Replace with actual API call once backend is available
-	
-	// Mock configuration for testing
-	mockConfig := map[string]interface{}{
-		"logging": map[string]interface{}{
-			"level":    "INFO",
-			"rotation": "daily",
-		},
-		"monitoring": map[string]interface{}{
-			"enabled":    true,
-			"prometheus": true,
-			"grafana":    true,
-		},
-		"security": map[string]interface{}{
-			"encryption": "AES256",
-			"tls_version": "1.3",
-		},
-	}
-	
-	config := &model.AmenitiesConfiguration{
-		OrganizationID:        organizationID,
-		Environment:          environment,
-		ConfigurationTemplate: mockConfig,
-		Version:              "1.0.0",
-	}
-	
-	return config, nil
-}
-
-// CheckDeploymentCellConfigurationDrift checks for configuration drift in deployment cells
-// This is a placeholder implementation - the actual API endpoint may not exist yet
-func CheckDeploymentCellConfigurationDrift(ctx context.Context, token string, deploymentCellID string, organizationID string, environment string) (*model.DeploymentCellAmenitiesStatus, error) {
-	// TODO: Replace with actual API call once backend is available
-	
-	// Mock drift detection results
-	driftDetails := []model.ConfigurationDrift{
-		{
-			Path:         "logging.level",
-			CurrentValue: "DEBUG",
-			TargetValue:  "INFO",
-			DriftType:    "different",
-		},
-		{
-			Path:         "monitoring.grafana",
-			CurrentValue: nil,
-			TargetValue:  true,
-			DriftType:    "missing",
-		},
-	}
-	
-	status := &model.DeploymentCellAmenitiesStatus{
-		DeploymentCellID:      deploymentCellID,
-		HasConfigurationDrift: len(driftDetails) > 0,
-		DriftDetails:          driftDetails,
-		HasPendingChanges:     false,
-		Status:               "drift_detected",
-		LastCheck:            time.Now(),
-	}
-	
-	return status, nil
-}
-
-// SyncDeploymentCellWithTemplate synchronizes deployment cell with organization template
-// This places changes in a pending state
-func SyncDeploymentCellWithTemplate(ctx context.Context, token string, deploymentCellID string, organizationID string, environment string) (*model.DeploymentCellAmenitiesStatus, error) {
-	// TODO: Replace with actual API call once backend is available
-	
-	// Mock pending changes based on drift
-	pendingChanges := []model.PendingConfigurationChange{
-		{
-			Path:      "logging.level",
-			Operation: "update",
-			OldValue:  "DEBUG",
-			NewValue:  "INFO",
-		},
-		{
-			Path:      "monitoring.grafana",
-			Operation: "add",
-			NewValue:  true,
-		},
-	}
-	
-	status := &model.DeploymentCellAmenitiesStatus{
-		DeploymentCellID:      deploymentCellID,
-		HasConfigurationDrift: false, // Drift resolved by sync
-		HasPendingChanges:     len(pendingChanges) > 0,
-		PendingChanges:        pendingChanges,
-		Status:               "pending_changes",
-		LastCheck:            time.Now(),
-	}
-	
-	return status, nil
-}
-
-// ApplyPendingChangesToDeploymentCell applies pending configuration changes to deployment cell
-// This uses the existing ApplyPendingChangesToHostCluster API from the SDK
-func ApplyPendingChangesToDeploymentCell(ctx context.Context, token string, serviceID string, environmentID string, deploymentCellID string) error {
-	ctxWithToken := context.WithValue(ctx, openapiclientfleet.ContextAccessToken, token)
-	apiClient := getFleetClient()
-
-	req := apiClient.InventoryApiAPI.InventoryApiApplyPendingChangesToHostCluster(ctxWithToken, serviceID, environmentID, deploymentCellID)
-
-	var r *http.Response
-	defer func() {
-		if r != nil {
-			_ = r.Body.Close()
-		}
-	}()
-
-	r, err := req.Execute()
-	if err != nil {
-		return handleFleetError(err)
-	}
-
-	return nil
-}
-
-// ListAvailableEnvironments lists available environments for amenities configuration
-// This is a placeholder implementation - the actual API endpoint may not exist yet
-func ListAvailableEnvironments(ctx context.Context, token string) ([]model.AmenitiesEnvironment, error) {
-	// TODO: Replace with actual API call once backend is available
-	
-	// Mock environments
-	environments := []model.AmenitiesEnvironment{
-		{
-			Name:        "production",
-			DisplayName: "Production",
-			Description: "Production environment configuration",
-		},
-		{
-			Name:        "staging",
-			DisplayName: "Staging",
-			Description: "Staging environment configuration",
-		},
-		{
-			Name:        "development",
-			DisplayName: "Development",
-			Description: "Development environment configuration",
-		},
-	}
-	
-	return environments, nil
-}
-
-// GetDeploymentCellAmenitiesStatus retrieves the current amenities status for a deployment cell
-// This is a placeholder implementation - the actual API endpoint may not exist yet
-func GetDeploymentCellAmenitiesStatus(ctx context.Context, token string, deploymentCellID string) (*model.DeploymentCellAmenitiesStatus, error) {
-	// TODO: Replace with actual API call once backend is available
-	
-	// Mock current status
-	status := &model.DeploymentCellAmenitiesStatus{
-		DeploymentCellID:      deploymentCellID,
-		HasConfigurationDrift: false,
-		HasPendingChanges:     false,
-		Status:               "synchronized",
-		LastCheck:            time.Now().Add(-1 * time.Hour),
-	}
-	
-	return status, nil
-}
-
-// ValidateAmenitiesConfiguration validates the provided amenities configuration
-func ValidateAmenitiesConfiguration(configTemplate map[string]interface{}) error {
-	if configTemplate == nil {
-		return fmt.Errorf("configuration template cannot be nil")
-	}
-	
-	if len(configTemplate) == 0 {
-		return fmt.Errorf("configuration template cannot be empty")
-	}
-	
-	// Add more validation logic as needed based on the actual configuration schema
-	
-	return nil
-}
