+++ mkdocs/docs/omnistrate-ctl_deployment-cell.md
@@ -27,9 +27,11 @@ omnistrate-ctl deployment-cell [operation] [flags]
 
 * [omnistrate-ctl](omnistrate-ctl.md)	 - Manage your Omnistrate SaaS from the command line
 * [omnistrate-ctl deployment-cell adopt](omnistrate-ctl_deployment-cell_adopt.md)	 - Adopt a deployment cell
-* [omnistrate-ctl deployment-cell amenities](omnistrate-ctl_deployment-cell_amenities.md)	 - Manage deployment cell amenities synchronization
+* [omnistrate-ctl deployment-cell apply](omnistrate-ctl_deployment-cell_apply.md)	 - Apply pending changes to deployment cell
+* [omnistrate-ctl deployment-cell check-drift](omnistrate-ctl_deployment-cell_check-drift.md)	 - Check deployment cell for configuration drift
 * [omnistrate-ctl deployment-cell delete](omnistrate-ctl_deployment-cell_delete.md)	 - Delete a deployment cell
 * [omnistrate-ctl deployment-cell list](omnistrate-ctl_deployment-cell_list.md)	 - List all deployment cells
 * [omnistrate-ctl deployment-cell status](omnistrate-ctl_deployment-cell_status.md)	 - Get status of a deployment cell
+* [omnistrate-ctl deployment-cell sync](omnistrate-ctl_deployment-cell_sync.md)	 - Sync deployment cell with organization+environment template
 * [omnistrate-ctl deployment-cell update-kubeconfig](omnistrate-ctl_deployment-cell_update-kubeconfig.md)	 - Update kubeconfig for a deployment cell
