+++ mkdocs/docs/omnistrate-ctl_service_init-amenities.md
@@ -1,22 +1,29 @@
-## omnistrate-ctl organization amenities init
+## omnistrate-ctl service init-amenities
 
-Initialize organization-level amenities configuration template
+Initialize service provider organization amenities configuration template
 
 ### Synopsis
 
-Initialize organization-level amenities configuration template through an interactive process.
+Initialize service provider organization-level amenities configuration template through an interactive process.
 
 This command starts an interactive process to define the default organization-level 
-amenities configuration template. This step is purely at the org level; no reference to any 
-service is needed.
+amenities configuration template. This step is purely at the service provider org level; 
+no reference to any specific service is needed.
 
 The configuration will be stored as a template that can be applied to different 
 environments (production, staging, development) and used to synchronize deployment cells.
 
 Organization ID is automatically determined from your credentials.
 
+Examples:
+  # Initialize amenities configuration interactively
+  omnistrate-ctl service init-amenities -e production
+
+  # Initialize from YAML file
+  omnistrate-ctl service init-amenities -e production -f sample-amenities.yaml
+
 ```
-omnistrate-ctl organization amenities init [flags]
+omnistrate-ctl service init-amenities [flags]
 ```
 
 ### Options
@@ -24,7 +31,7 @@ omnistrate-ctl organization amenities init [flags]
 ```
   -f, --config-file string   Path to configuration YAML file (optional)
   -e, --environment string   Target environment (production, staging, development)
-  -h, --help                 help for init
+  -h, --help                 help for init-amenities
       --interactive          Use interactive mode to configure amenities (default true)
 ```
 
@@ -37,5 +44,5 @@ omnistrate-ctl organization amenities init [flags]
 
 ### SEE ALSO
 
-* [omnistrate-ctl organization amenities](omnistrate-ctl_organization_amenities.md)	 - Manage organization amenities configuration templates
+* [omnistrate-ctl service](omnistrate-ctl_service.md)	 - Manage Services for your account
